+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2023-03-26
+## [1.1.1] - 2023-03-29
+### Added
+- Added `NewNotFoundError`.
+- Added built-in languages constants.
+- Now when using `WithLanguage`, it lookups for both `xy` and `xy-ZW`.
+- Added translations to the prefix.
+
+### Changed
+- Improved error messages.
+- ## [1.1.0] - 2023-03-26
 ### Changed
 - Added Catalog of errors.
 - Added support for translations.